+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Data Manager query string management to merge instead of replace
 
+### Fixed
+
+- Summary and Calculation Elements not showing on read only forms
+
 ## [2.5.2] - 2023-01-16
 
 ### Fixed